+++ iris/service/content/file/__init__.py
@@ -2,4 +2,11 @@ from .document import File  # noqa
 
 
 def includeme(config):
-    pass
+    from iris.service.rest import auth
+    config.add_route('file_public_api',
+                     'files',
+                     static=True)
+    config.add_route('file_admin_api',
+                     'admin/files',
+                     static=True,
+                     factory=auth.AdminServiceAuthFactory)