+++ iris/service/content/file/document.py
@@ -1,13 +1,20 @@
 from lovely.esdb.document import Document
-from lovely.esdb.properties import Property
+from lovely.esdb.properties import Property, LocalRelation
 
 from iris.service.db.dc import dc_defaults, DC_CREATED, DC_MODIFIED
+from .tempstorage import get_temp_upload_path
+
+
+class StorageType(object):
+    S3 = "s3"
+    TMP = "tmp"
 
 
 class File(Document):
-    """A file in the database
+    """A file in the database.
 
-    TODO: needs to be completed
+    This is a meta data container for the file. The file remains in an external
+    storage.
     """
 
     INDEX = 'files'
@@ -18,3 +25,46 @@ class File(Document):
         default=dc_defaults(DC_CREATED, DC_MODIFIED),
         doc="Dublin Core data."
     )
+
+    state = Property(
+        doc="The state of the file (visible/hidden)"
+    )
+
+    original_name = Property(
+        doc="The original file name before upload"
+    )
+
+    _owner_id = Property(
+        name="owner_id",
+        doc="The ID of the user/session user who uploaded the file"
+    )
+    owner = LocalRelation('_owner_id', 'User.id')
+
+    storage_type = Property(
+        doc="""The kind of storage that has been used for the file.
+
+        `s3` for files stored on S3, `tmp` for files stored in a temporary
+        folder (used for local/testing environments).
+        """
+    )
+
+    content_type = Property(
+        doc="The file's guessed MIME type (e.g. `text/plain`, `image/jpeg`)"
+    )
+
+    @property
+    def url(self):
+        """The download URL to retrieve the file.
+        """
+        if self.storage_type == StorageType.S3:
+            from .s3 import get_s3_url
+            return get_s3_url(self.id)
+        elif self.storage_type == StorageType.TMP:
+            # for local environment or testing
+            return "file://%s/%s" % (get_temp_upload_path(), self.id)
+        return None
+
+    def get_source(self):
+        res = super(File, self).get_source()
+        res['url'] = self.url
+        return res