+++ CHANGES.rst
@@ -5,6 +5,10 @@ Changes For IRIS Service
 unreleased
 ==========
 
+ - added file service (upload and download)
+   DEPLOY:
+    - recreate files table in Crate
+    - enable libmagic in Docker containers
  - provide a reason in the support endpoint if the token is invalid
  - handle SMS send exceptions with a ValueError